+++ src/main/java/com/example/shop/repository/CarouselItemRepository.java
@@ -9,7 +9,7 @@ import com.example.shop.entity.Authorization;
 import com.example.shop.entity.CarouselItem;
 import com.example.shop.entity.CarouselView;
 @RepositoryRestResource
-public interface CarouselItemRepository extends JpaRepository<CarouselItem, Integer> {
+public interface CarouselItemRepository extends JpaRepository<CarouselItem, Long> {
 	List<CarouselItem> findByCarouselView(CarouselView cv);
 	List<CarouselItem> findByCarouselView(Integer car_id);
