+++ src/main/java/com/example/shop/repository/GridItemRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import com.example.shop.entity.GridItem;
 import com.example.shop.entity.GridView;
 @RepositoryRestResource
-public interface GridItemRepository extends JpaRepository<GridItem, Integer> {
+public interface GridItemRepository extends JpaRepository<GridItem, Long> {
 
 	
 	List<GridItem> findByGridView(GridView grid);
